feat: Content view의 기본적인 모양 설정
