--- ProjectManager/ProjectManager/ContentView.swift
@@ -1,20 +1,0 @@
-//
-//  ContentView.swift
-//  ProjectManager
-//
-//  Created by JINHONG AN on 2021/10/26.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
